+++ packages/components/package.json
@@ -1,9 +1,9 @@
 {
   "name": "@ds-pack/components",
   "version": "1.0.11",
-  "main": "dist/index.es5.js",
-  "module": "dist/index.mjs",
+  "main": "dist/index.js",
   "source": "src/index.ts",
+  "types": "dist/index.d.ts",
   "repository": "git@github.com:ds-pack/components.git",
   "author": "Matt Hamlin <matthewjameshamlin@gmail.com>",
   "license": "MIT",
@@ -11,7 +11,7 @@
     "build-styles": "rollup -c rollup.config.mjs",
     "build-js": "babel --extensions .ts,.tsx ./src --out-dir ./dist --ignore './src/**/*.test.tsx,./src/**/*.test.ts'",
     "build-ts": "tsc --emitDeclarationOnly --outDir ./dist --project tsconfig.build.json",
-    "build": "yarn build-styles && yarn build-js && yarn build-ts",
+    "build": "rm -rf ./dist && yarn build-styles && yarn build-js && yarn build-ts",
     "test": "jest",
     "test-watch": "jest --watch",
     "test-cov": "jest --collectCoverage",
@@ -38,10 +38,8 @@
     "@vanilla-extract/babel-plugin": "^1.1.4",
     "@vanilla-extract/css": "^1.6.8",
     "@vanilla-extract/dynamic": "^2.0.2",
-    "@vanilla-extract/esbuild-plugin": "^2.0.2",
     "@vanilla-extract/rollup-plugin": "^1.2.0",
     "babel-jest": "^26",
-    "esbuild": "^0.14.25",
     "fast-glob": "^3.2.12",
     "jest": "^26",
     "nanopop": "^2",
@@ -60,7 +58,6 @@
   },
   "peerDependencies": {
     "@ds-pack/use-refs": "*",
-    "@vanilla-extract/css": "^1.6.8",
     "@vanilla-extract/dynamic": "^2.0.2",
     "nanopop": "^2",
     "rainbow-sprinkles": "^0.12",
