+++ packages/components/src/ListItem.ts
@@ -1,0 +1,3 @@
+import { ListItem } from './List'
+
+export { ListItem }
