+++ packages/components/src/Portal/Portal.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useEffect, useLayoutEffect, useState } from 'react'
 import { createPortal } from 'react-dom'
