+++ packages/components/src/themeClass.ts
@@ -1,0 +1,3 @@
+import { themeClass } from './vars.css'
+
+export { themeClass }
