+++ sandbox/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "typescript.tsdk": "../node_modules/typescript/lib",
+  "typescript.enablePromptUseWorkspaceTsdk": true
+}
