+++ sandbox/app/layout.tsx
@@ -1,0 +1,14 @@
+import { themeClass } from '@ds-pack/components'
+import '@ds-pack/components/dist/reset.css'
+
+export default function Layout({ children }) {
+  return (
+    <html className={themeClass} lang="en-US">
+      <head>
+        <meta name="viewport" content="initial-scale=1.0, width=device-width" />
+        <title>Sandbox</title>
+      </head>
+      <body>{children}</body>
+    </html>
+  )
+}
