+++ sandbox/next.config.js
@@ -1,8 +1,15 @@
 /** @type {import('next').NextConfig} */
 let config = {
   reactStrictMode: true,
+  modularizeImports: {
+    '@ds-pack/components': {
+      transform: '@ds-pack/components/dist/{{member}}',
+      skipDefaultConversion: true,
+    },
+  },
   experimental: {
     runtime: 'nodejs',
+    appDir: true,
   },
 }
