--- sandbox/src/pages/_document.tsx
@@ -1,14 +1,0 @@
-import { themeClass } from '@ds-pack/components'
-import { Html, Head, Main, NextScript } from 'next/document'
-
-export default function Document() {
-  return (
-    <Html className={themeClass}>
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
